requirements.txt
numpy
pandas
matplotlib
seaborn
spotipy
scikit-learn
torch
cv2
torchvision
torchaudio
torchtext
streamlit
ollama
opencv-python
transformers
dorothy-cci
datasets
tiktoken
onnxruntime
Pillow
imageio
sounddevice
librosa
ultralytics
scikit-image
acids-rave
open_clip_torch
streamlit-drawable-canvas
scikit-learn
Flask
requests
python-dotenv
spotipy